chore: Improve code organization
