+++ src/sinks/assert/prelude.rs
@@ -1,0 +1,24 @@
+use crate::model::{BlockRecord, Event};
+
+#[derive(Default, Debug)]
+pub(crate) struct State {
+    pub current_event: Option<Event>,
+    pub previous_event: Option<Event>,
+    pub current_block: Option<BlockRecord>,
+    pub previous_block: Option<BlockRecord>,
+}
+
+pub(crate) enum Outcome {
+    Pass,
+    Fail,
+    Unknown,
+}
+
+impl From<bool> for Outcome {
+    fn from(other: bool) -> Self {
+        match other {
+            true => Outcome::Pass,
+            false => Outcome::Fail,
+        }
+    }
+}
